Refactor AddButton import paths for consistency across components
