--- src/components/screens/notes/index.tsx
@@ -1,2 +1,0 @@
-export * from './AddButton';
-export * from './NotesList';
